README.md
README
This README provides important information and a disclaimer regarding the code shared in this repository. It emphasizes that the author cannot be held responsible for any misuse or malicious activities conducted with the code. Users are urged to read and understand the terms, use the code at their own risk, and comply with legal and ethical guidelines. The disclaimer highlights the need for responsible and lawful utilization of the code, while acknowledging that users assume full responsibility for their actions.